requirements.txt
django-bootstrap-v5
pylint